Convert to use a Range to avoid expensive ops.
This avoids cloning `required_indent` into `state`, which used to happen
for each line that we needed to indent.

This also avoids draining the working Vec in `state` to remove N bytes
from the front. That's an expensive operation for a Vec since all
remaining elements need to be moved up.

Using a Range means we basically just increment 1 integer, the `start`
field, as we progress with writing the indentation.

In practice, it is likely that the entire `required_indent` will be short
and hence get written in one call anyway so the whole operation will boil
down to 1 call to the Range constructor and 1 call to Range.clone() - a
couple of usizes.