Remove the lifetime parameter on IndentWriter.
This simplifies the life of users. It is not egregious for the writer
to take ownership of the indent string, especially now that we are
using the inc()/dec() model, as there will actually be fewer allocations
than when using the old model - in the old model, to do increasing
indentation you had to allocate an entire new writer to wrap the first
one. In the new model you just increment the indent_level integer and
you only need to allocate 1 writer in the common case of using the same
indentation string at each level of indentation.